Implement Basic UI Component Structure as per Figma Design
