+++ src/App.jsx
@@ -1,7 +1,20 @@
+/** @format */
+
 import React from 'react'
+import Accounts from './components/Accounts'
+import ButtonSection from './components/ButtonSection'
+import Header from './components/Header'
+import Spendings from './components/Spendings'
 
 function App() {
-  return <div>Hello World!</div>
+    return (
+        <>
+            <Header />
+            <ButtonSection />
+            <Accounts />
+            <Spendings />
+        </>
+    )
 }
 
 export default App
