+++ src/components/ButtonSection.jsx
@@ -1,0 +1,14 @@
+/** @format */
+
+import React from 'react'
+
+function ButtonSection() {
+    return (
+        <div className="button-section">
+            <button className="button-section__button">Pay</button>
+            <button className="button-section__button">Transfer</button>
+        </div>
+    )
+}
+
+export default ButtonSection
