+++ src/components/Header.jsx
@@ -1,0 +1,15 @@
+/** @format */
+
+import React from 'react'
+import NavBar from './NavBar'
+
+function Header() {
+    return (
+        <header>
+            <h1>Best Bank</h1>
+            <NavBar />
+        </header>
+    )
+}
+
+export default Header
