+++ src/components/NavBar.jsx
@@ -1,0 +1,13 @@
+/** @format */
+
+import React from 'react'
+
+function NavBar() {
+    return (
+        <nav>
+            <h1>NavBar</h1>
+        </nav>
+    )
+}
+
+export default NavBar
