+++ src/components/Spendings.jsx
@@ -1,0 +1,9 @@
+/** @format */
+
+import React from 'react'
+
+function Spendings() {
+    return <h1>Spendings</h1>
+}
+
+export default Spendings
